# AIdocs/specs/table-intelligence/04_system/openapi/table-intelligence-openapi-draft.yaml
...
openapi: 3.1.0
info:
  title: Table Intelligence API
  version: 0.1.8-draft
  description: |
    表解析パイプラインの **API 叩き台**（OpenAPI 3.1）。

    **正本**
    - 契約・振る舞い: [SPEC-TI-014](../SPEC-TI-014-api.md)
    - DTO・列挙・識別子: [SPEC-TI-006](../../01_foundation/SPEC-TI-006-io-data.md)

    **ID / ref（レイヤ差）**
    - `evaluation_ref`（API・Job 応答）と DB 主キー `evaluation_id` は **同一キー空間**（015 §5.2）。
    - `suggestion_run_ref`（API・Job 応答）と DB 主キー `suggestion_run_id` は **同一キー空間**。
    - 人確認は **`session_id` のみ**を用いる（`review_session_id` は採用しない）。

    **責務**
    - `decision`: 002 Judgment（一次判定）。
    - `decision_recommendation`: 011 の推奨（002 を上書きしない）。
    - suppression の正本: 005。suggestion で suppression 適用済みを返す場合も **005 の状態を読んだ結果**である（本 API は窓口）。

    TODO: 認証スキーム詳細、ページネーション、012 の細粒度 `error_code` マトリクスは次版で確定。

    **MVP 実装整合（0.1.1）**
    - Django バックエンド実装に合わせ、Jobs の受理形・`TiMvpJobSummary` / `TiMvpJobDetail`、suppression GET の配列レスポンス、rerun の **201**、冪等 **200** をドラフトに反映した。
    - 006 完全準拠の `JobKind` / `HumanReviewSession.table_id` 等は **Phase 4** で本スキーマと実装の双方を厳密化する。
    - **クライアント向け**: 何に依存してよいか／Phase 4 で何が変わるかは **SPEC-TI-014 §18** を正とする。
    - **P4-2（0.1.2）**: `GET /datasets/{id}`, `GET /metadata/{id}`, `GET /evaluations/{ref}`, `GET /review-sessions/{id}` および session **POST 201** のレスポンスを **`TiMvp*` schema** に明示紐付け。006 理想形（`NormalizedDataset` 等）は **Phase 4 目標**として components に残す。
...

  license:
    name: Proprietary
servers:
  - url: https://api.example.com/v1
    description: Draft base URL（実装で置換）
security:
  - tokenAuth: []
...
paths:
  /table-analysis/jobs:
    post:
      tags: [Jobs]
      operationId: startAnalysisJob
      summary: 表解析ジョブ開始
      description: |
        原本参照を受理し非同期実行。レスポンスは **受理と `job_id`** を主とし、成果物本体は GET で再取得（014 §4.2）。

        **冪等**: ヘッダ `Idempotency-Key`（最大 128 文字）が同一 workspace で再送された場合、**新規行を増やさず**既存ジョブを返す。HTTP は **200**（MVP 実装）。
        新規受理時は **202**。

        **エラー**: 未知・無権限の `workspace_id` は **404 マスク**（越境判別不可）。
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
          description: 同一意味の POST の冪等キー（014 §4.5）。workspace スコープと組み合わせて一意。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartAnalysisJobRequest"
      responses:
        "200":
          description: 冪等リプレイ。既存 `job_id` を返し、重複投入しない（MVP）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpJobSummary"
        "202":
          description: 新規ジョブを受理。処理は非同期（テスト用設定ではインライン完了しうる）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpJobSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 未知の `workspace_id`（テナント分離・列挙攻撃対策で 404 マスクしうる。014 §14.3）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /table-analysis/jobs/{job_id}:
    get:
      tags: [Jobs]
      operationId: getAnalysisJob
      summary: ジョブ状態取得
      description: |
        確定済み `table_id` / ref の部分集合を返しうる。
        存在しない `job_id` および **他 workspace のジョブ**は **404 マスク**（同一応答）。
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: OK（MVP は `TiMvpJobDetail`。確定 ref は `artifact_refs` 束に集約）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpJobDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /table-analysis/jobs/{job_id}/rerun:
    post:
      tags: [Jobs]
      operationId: rerunAnalysisJob
      summary: 単一段階 rerun
      description: |
        001〜004 のいずれかから再実行。新 `job_id` ないし成果物世代が発生しうる（014 §12）。
        元 `job_id` が存在しない・越境の場合は **404**。**409**（競合・stale）は将来 014 §12 / §13 で拡張しうるが **MVP では主に未使用**。
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RerunJobRequest"
      responses:
        "201":
          description: |
            新 `job_id` のジョブ行を **Created** として返す（MVP・REST 慣習）。パイプライン非同期投入は実装フラグに依存。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpJobSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
...
  /datasets/{dataset_id}:
    get:
      tags: [Datasets]
      operationId: getNormalizedDataset
      summary: NormalizedDataset 取得（MVP 事実形）
      description: |
        **実装契約（0.1.2）**: レスポンスは `TiMvpNormalizedDataset`。
        `NormalizedDataset`（006 理想形・トップレベル `rows` / `trace_map`）は Phase 4 で **additive** 導入予定。現行は `dataset_payload` に集約。
      parameters:
        - $ref: "#/components/parameters/DatasetId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpNormalizedDataset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /metadata/{metadata_id}:
    get:
      tags: [Metadata]
      operationId: getAnalysisMetadata
      summary: AnalysisMetadata 取得（MVP 事実形）
      description: |
        **実装契約（0.1.2）**: レスポンスは `TiMvpAnalysisMetadata`。
        `AnalysisMetadata`（006 理想形・例: `grain` 必須）は Phase 4 目標。現行は `dimensions` / `measures` / `time_axis` / `decision` 等の MVP 投影。
      parameters:
        - $ref: "#/components/parameters/MetadataId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpAnalysisMetadata"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /metadata/{metadata_id}/review-points:
    get:
      tags: [Metadata]
      operationId: listReviewPoints
      summary: review_points 取得
      description: 004 正本の `review_points[]` の投影。
      parameters:
        - $ref: "#/components/parameters/MetadataId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [review_points]
                properties:
                  metadata_id:
                    type: string
                  review_points:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
...
  /review-sessions:
    post:
      tags: [ReviewSessions]
      operationId: createReviewSession
      summary: review session 作成
      description: |
        参照する `metadata_id` 等が存在しない・越境の場合は **404 マスク**（014 §14.3）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewSessionRequest"
      responses:
        "201":
          description: 作成（MVP 事実形 `TiMvpHumanReviewSession`）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpHumanReviewSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /review-sessions/{session_id}:
    get:
      tags: [ReviewSessions]
      operationId: getReviewSession
      summary: review session 取得（MVP 事実形）
      description: |
        **実装契約（0.1.2）**: レスポンスは `TiMvpHumanReviewSession`。
        `HumanReviewSession`（006 理想形・`table_id` / `pending_questions`）は Phase 4 目標。
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpHumanReviewSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /review-sessions/{session_id}/answers:
    post:
      tags: [ReviewSessions]
      operationId: submitReviewAnswers
      summary: answers 登録
      description: |
        **005 MVP 事実形**（Django `SubmitReviewAnswersRequestSerializer`）。**Request**: `TiMvpSubmitReviewAnswersRequest`。
        各回答は **`question_key` + `answer_value`（任意 JSON）** が契約。006 の `point_id` / `client_nonce` / `ReviewAnswer` 完全形とは **未整合**（Phase 4 **additive**）。
        任意 **`mark_resolved`**: `true` のとき session を **RESOLVED** に。**`resolution_grade`**: 任意文字列（最大 64、空・null 可）を session に保存。
        存在しない `session_id`・越境は **404 マスク**。**409** は stale / 二重送信等の **将来枠**（MVP では多く未使用）。
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TiMvpSubmitReviewAnswersRequest"
      responses:
        "200":
          description: |
            **HTTP 200**（実装事実）。`session`（`TiMvpHumanReviewSession`）と `answers`（**本リクエストで作成した行**の `TiMvpHumanReviewAnswer[]`）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpSubmitReviewAnswersResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: stale / 競合（将来。014 §13）。MVP では主に未使用。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /review-sessions/{session_id}/rerun:
    post:
      tags: [ReviewSessions]
      operationId: rerunAfterReview
      summary: review 後 rerun
      description: |
        005 正本の遷移に従い再パイプラインを起動。
        存在しない `session_id`・越境は **404 マスク**。**409** は将来の状態競合用（MVP では主に未使用）。
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "202":
          description: 上流再実行ジョブを受理（MVP は `TiMvpJobSummary`）。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TiMvpJobSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 状態競合（将来）。MVP では主に未使用。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
...
  /suggestion-runs:
    post:
      tags: [Suggestions]
      operationId: startSuggestionRun
      summary: suggestion run 開始
      description: |
        **主入力は `metadata_id`（必須寄り）**。013 は 004 を主入力とし、011/005 を読む側。
        参照 `metadata_id` が存在しない・越境の場合は **404 マスク**。
        **409**: 当該 `metadata_id` が **既に superseded**（`artifact_relation` に **旧 `analysis_metadata` → 新 `analysis_metadata`** の `TI_JOB_RERUN_SUPERSEDES` または `TI_REVIEW_RERUN_SUPERSEDES` がある）。**最新の metadata_id** で再試行（014 §13.1）。**`is_latest` 列は不要**（edge 有無で判定）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSuggestionRunRequest"
      responses:
        "202":
          content:
            application/json:
              schema:
                type: object
                required: [suggestion_run_ref]
                properties:
                  suggestion_run_ref:
                    type: string
                    description: 006 `SuggestionSet.suggestion_run_id` と同一キー空間
                  job_id:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            **stale / superseded metadata**: `artifact_relation` に当該 ID が **from** 側の metadata SUPERSEDES edge がある。`error_code` は **`TI_CONFLICT`**。本文の `detail` に再取得のヒント。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /suggestion-runs/{suggestion_run_ref}:
    get:
      tags: [Suggestions]
      operationId: getSuggestionSet
      summary: SuggestionSet 取得
      description: |
        存在しない ref・越境は **404 マスク**。
      parameters:
        - $ref: "#/components/parameters/SuggestionRunRef"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
  /suggestion-runs/{suggestion_run_ref}/candidates:
    get:
      tags: [Suggestions]
      operationId: listSuggestionCandidates
      summary: 候補一覧 / 条件付き取得
      description: |
        - `suppression=applied`: 005 の suppression を反映した結果であることを前提とする（クエリは実装で確定）。
        - `include=recommendation`: 011 由来の `decision_recommendation` 等を同梱しうる（002 `decision` と混在させない）。
        - 存在しない ref・越境は **404 マスク**。
        - **409 は付けない**（読取は成功時 **200** のみ。stale は **`POST /suggestion-runs`** 側で検知—014 §13.1）。
      parameters:
        - $ref: "#/components/parameters/SuggestionRunRef"
        - name: suppression
          in: query
          schema:
            type: string
            enum: [applied, none]
          description: TODO 014 次版で enum 確定
        - name: include
          in: query
          schema:
            type: string
            enum: [recommendation]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [candidates]
                properties:
                  candidates:
                    type: array
                    items:
                      $ref: "#/components/schemas/SuggestionCandidate"
                  decision_recommendation:
                    $ref: "#/components/schemas/DecisionRecommendation"
                    description: include=recommendation 時のみ。011 正本。
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/ErrorResponseDefault"
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        MVP 既定は DRF **Token** 認証（`Authorization: Token <key>`）。Session 等は実装・環境により異なりうる。
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
...
    DatasetId:
      name: dataset_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: |
        入力検証エラー。`/api/v1/*` では `error_code` は多くの場合 **`TI_VALIDATION_ERROR`**（フィールドエラーは `errors`）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              summary: フィールド検証
              value:
                error_code: TI_VALIDATION_ERROR
                detail: Validation failed.
                errors:
                  workspace_id:
                    - This field is required.
    NotFound:
      description: |
        リソース不明または **workspace 越境の 404 マスク**（014 §14.3）。`error_code` は **`TI_NOT_FOUND`**。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error_code: TI_NOT_FOUND
                detail: Not found.
    Unauthorized:
      description: |
        未認証。`error_code` は **`TI_AUTHENTICATION_REQUIRED`**（DRF 設定により 401）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ErrorResponseDefault:
      description: その他エラー。`error_code` は **`TI_ERROR`** 等。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      description: |
        **table_intelligence（`/api/v1/*`）** のエラー **事実契約**（P4-1 / SPEC-TI-012 最小寄せ）。
        DRF 既定の本文を **ラップ**し、**`detail` を維持**しつつ **`error_code` を必須**とする。Phase 4 で `message` / `details` / `request_id` を **additive** 拡張しうる。
        **互換**: 既存クライアントは引き続き `detail` を読める。`error_code` で分岐可能。
      required: [error_code, detail]
      properties:
        error_code:
          type: string
          description: |
            実装が付与する安定キー（`TI_` 接頭辞）。代表: `TI_NOT_FOUND`, `TI_VALIDATION_ERROR`,
            `TI_BAD_REQUEST`, `TI_AUTHENTICATION_REQUIRED`, `TI_PERMISSION_DENIED`, `TI_CONFLICT`, `TI_ERROR`
        detail:
          description: 人向け要約または DRF 由来の `detail`（文字列・配列・オブジェクトのいずれもありうる）
          oneOf:
            - type: string
            - type: object
            - type: array
        errors:
          type: object
          additionalProperties: true
          description: |
            主に HTTP 400 のフィールド検証。キーはフィールド名、値はメッセージの配列（DRF 慣習）。
            `detail` のみで足りる場合は省略。
        message:
          type: string
          description: 012 拡張・英語要約用（省略可。後続 Phase 4）
        details:
          type: object
          additionalProperties: true
          description: 012 構造化補足（省略可。後続 Phase 4）
    StartAnalysisJobRequest:
      type: object
      description: |
        **MVP 実装**は `workspace_id` 必須、`source_type` / `source_ref` / `request_payload` 任意。
        OpenAPI 上の `source_id` / `document_id` は 006・014 次版向け; 実装では `source_ref` 等に寄せうる。
      required: [workspace_id]
      properties:
        workspace_id:
          type: string
        source_type:
          type: string
          description: 原本種別（実装 MVP の文字列ラベル）
        source_ref:
          type: string
          description: 原本参照（URI 等）
        request_payload:
          type: object
          additionalProperties: true
          description: 任意 JSON。rerun lineage や `mvp_pipeline` スナップショットが入りうる（015 `analysis_job` JSON 候補と整合）。
        source_id:
          type: string
          nullable: true
          description: 006 向け予約（MVP 実装未使用の場合あり）
        document_id:
          type: string
          nullable: true
          description: 006 向け予約（MVP 実装未使用の場合あり）
        idempotency_key:
          type: string
          nullable: true
          deprecated: true
          description: ボディでの冪等キー指定は **非推奨**。ヘッダ `Idempotency-Key` を正とする（MVP）。
...
    TiMvpJobSummary:
      type: object
      description: |
        Django MVP のジョブ受理・一覧サマリ。006 `JobKind` 列は **未搭載**のため `kind` は含めない。
        `artifact_refs` は未生成時 **null**、確定後は主要 ID の束（014 の部分集合）。
      required: [job_id, workspace_id, status, current_stage, created_at, artifact_refs]
      properties:
        job_id:
          type: string
        workspace_id:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        current_stage:
          type: string
          description: 実装固有段階（例 `queued`, `rerun_requested`, `mvp_pipeline_materialized`）。
        created_at:
          type: string
          format: date-time
        artifact_refs:
          oneOf:
            - $ref: "#/components/schemas/ArtifactRefs"
            - type: "null"
...